requirements.txt
django
djangorestframework
djangorestframework-gis
psycopg2-binary
Pillow
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
earthengine-api
django-cors-headers